Try to increase unzip performance
Trying to increase the performance of the unzip function. Currently it will not set the file headers after unzipping
as this is having a severe impact on the performance. The `progressMonitor` seems to have little to no impact at all
on the performance, which is better than initially expected.

By moving to `BufferedOutputStream` in combination with `InputStream.transferTo()` method performance got a 25% boost.
This is however a Java 9 function, so the whole package is no longer backwards compatible with Java 8 or 7.

Need to further investigate the file header issue, because this should not have a negative impact at all.